+++ azure_monitor/tests/auto_collection/test_request_metrics.py
@@ -11,6 +11,7 @@ from azure_monitor.sdk.auto_collection import request_metrics
 from azure_monitor.sdk.auto_collection.metrics_span_processor import (
     AzureMetricsSpanProcessor,
 )
+from azure_monitor.sdk.auto_collection.utils import AutoCollectionType
 
 
 # pylint: disable=protected-access
@@ -35,19 +36,13 @@ class TestRequestMetrics(unittest.TestCase):
             meter=mock_meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         self.assertEqual(request_metrics_collector._meter, mock_meter)
         self.assertEqual(request_metrics_collector._labels, self._test_labels)
-        self.assertEqual(mock_meter.register_observer.call_count, 3)
+        self.assertEqual(mock_meter.register_observer.call_count, 2)
         create_metric_calls = mock_meter.register_observer.call_args_list
         create_metric_calls[0].assert_called_with(
-            callback=request_metrics_collector._track_request_failed_rate,
-            name="\\ApplicationInsights\\Requests Failed/Sec",
-            description="Incoming Requests Failed Rate",
-            unit="rps",
-            value_type=float,
-        )
-        create_metric_calls[1].assert_called_with(
             callback=request_metrics_collector._track_request_duration,
             name="\\ASP.NET Applications(??APP_W3SVC_PROC??)\\Request Execution Time",
             description="Incoming Requests Average Execution Time",
@@ -55,7 +50,7 @@ class TestRequestMetrics(unittest.TestCase):
             value_type=int,
         )
 
-        create_metric_calls[2].assert_called_with(
+        create_metric_calls[1].assert_called_with(
             callback=request_metrics_collector._track_request_rate,
             name="\\ASP.NET Applications(??APP_W3SVC_PROC??)\\Requests/Sec",
             description="Incoming Requests Rate",
@@ -68,6 +63,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         self._span_processor.request_duration = 100
         self._span_processor.request_count = 10
@@ -90,6 +86,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         self._span_processor.request_duration = 100
         self._span_processor.request_count = 10
@@ -113,6 +110,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         time_mock.time.return_value = 100
         request_metrics.requests_map["last_time"] = 98
@@ -137,6 +135,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         request_metrics.requests_map["last_time"] = None
         obs = Observer(
@@ -158,6 +157,7 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
             labels=self._test_labels,
             span_processor=self._span_processor,
+            collection_type=AutoCollectionType.STANDARD_METRICS,
         )
         time_mock.time.return_value = 100
         request_metrics.requests_map["last_rate"] = 5.0
@@ -171,74 +171,6 @@ class TestRequestMetrics(unittest.TestCase):
             meter=self._meter,
         )
         request_metrics_collector._track_request_rate(obs)
-        self.assertEqual(
-            obs.aggregators[tuple(self._test_labels.items())].current, 5.0
-        )
-
-    @mock.patch("azure_monitor.sdk.auto_collection.request_metrics.time")
-    def test_track_request_failed_rate(self, time_mock):
-        request_metrics_collector = request_metrics.RequestMetrics(
-            meter=self._meter,
-            labels=self._test_labels,
-            span_processor=self._span_processor,
-        )
-        time_mock.time.return_value = 100
-        request_metrics.requests_map["last_time"] = 98
-        self._span_processor.failed_request_count = 4
-        obs = Observer(
-            callback=request_metrics_collector._track_request_failed_rate,
-            name="test",
-            description="test",
-            unit="test",
-            value_type=float,
-            meter=self._meter,
-        )
-        request_metrics_collector._track_request_failed_rate(obs)
-        self.assertEqual(
-            obs.aggregators[tuple(self._test_labels.items())].current, 2.0
-        )
-
-    @mock.patch("azure_monitor.sdk.auto_collection.request_metrics.time")
-    def test_track_request_failed_rate_time_none(self, time_mock):
-        time_mock.time.return_value = 100
-        request_metrics_collector = request_metrics.RequestMetrics(
-            meter=self._meter,
-            labels=self._test_labels,
-            span_processor=self._span_processor,
-        )
-        request_metrics.requests_map["last_time"] = None
-        obs = Observer(
-            callback=request_metrics_collector._track_request_failed_rate,
-            name="test",
-            description="test",
-            unit="test",
-            value_type=float,
-            meter=self._meter,
-        )
-        request_metrics_collector._track_request_failed_rate(obs)
-        self.assertEqual(
-            obs.aggregators[tuple(self._test_labels.items())].current, 0.0
-        )
-
-    @mock.patch("azure_monitor.sdk.auto_collection.request_metrics.time")
-    def test_track_request_failed_rate_error(self, time_mock):
-        request_metrics_collector = request_metrics.RequestMetrics(
-            meter=self._meter,
-            labels=self._test_labels,
-            span_processor=self._span_processor,
-        )
-        time_mock.time.return_value = 100
-        request_metrics.requests_map["last_rate"] = 5.0
-        request_metrics.requests_map["last_time"] = 100
-        obs = Observer(
-            callback=request_metrics_collector._track_request_failed_rate,
-            name="test",
-            description="test",
-            unit="test",
-            value_type=float,
-            meter=self._meter,
-        )
-        request_metrics_collector._track_request_failed_rate(obs)
         self.assertEqual(
             obs.aggregators[tuple(self._test_labels.items())].current, 5.0
         )